fix tests

+++ src/users.py
@@ -88,9 +88,14 @@ def compute_total_period():
 
 
 class Users:
-    """Contains information about users."""
-    def __init__(self, path: str):
+    """
+    Contains information about users.
+    `path`: where the information about users is stored.
+    If `read_only` is set, don't write to `path`.
+    """
+    def __init__(self, path: str, read_only: bool = False):
         self.path = path
+        self.read_only = read_only
         self.global_lock = threading.Lock()
 
         self.read()
@@ -108,13 +113,16 @@ class Users:
                     break
 
         self.done = False
-        self.write_thread = threading.Thread(target=write_loop)
-        self.write_thread.start()
+        if not self.read_only:
+            self.write_thread = threading.Thread(target=write_loop)
+            self.write_thread.start()
 
     def finish(self):
         self.done = True
-        self.write_thread.join()
-        self.write()
+        if not self.read_only:
+            self.write_thread.join()
+            if self.dirty:
+                self.write()
 
     def read(self):
         with self.global_lock:
@@ -191,8 +199,9 @@ class Users:
             for user in self.users:
                 raw_users.append(user.as_dict())
 
-            print(f'Writing {len(self.users)} users to {self.path}')
-            with open(self.path, 'w') as f:
-                for raw_user in raw_users:
-                    print(json.dumps(raw_user), file=f)
+            if not self.read_only:
+                print(f'Writing {len(self.users)} users to {self.path}')
+                with open(self.path, 'w') as f:
+                    for raw_user in raw_users:
+                        print(json.dumps(raw_user), file=f)
             self.dirty = False
